Split some OSPF logic from VirtualRouter
